Error detection and handling added
This version improves the reliability of RS-bus software by including error detection and correction. It can detect and react upon parity errors and pulse count errors, and includes forward error correction options.
The default version for the MegaCoreX and DxCore has been changed, at the expense of somewhat higher CPU load but with the ability to detect spikes on the input pin.