Usage of Timer 5 on ATMega 2560
For ATMega 2560 processors we will use Timer 5 to call every 2ms  checkPolling() / updateAddressPolled(). Even on systems with code that can't be interrupted but take more than 2ms, the silence period will be detected reliable and the RS-Bus address being polled will be reset. This is particularly useful in combination with LCD display libraries, that easily take 5ms.