Added the SW with Timer1 variant
An additional RSBUS_USES_SW_T1 variant was added for "traditional" AtMega processors. Although this variant will be most interesting for processors that support several 16 bit timers, it can also be used for AtMega processors that support only a single Timer 16.
In fact this variant was added since (at the time if writing) MightyCore had problems with the millis() function on (AtMega16) boards with F_CPU of 11.0592Mhz. Maybe in future releases of this RS-Bus library we will therefore replace all millis() calls with micros() calls.