Replaced millis() by micros()
Replaced everywhere the check "if ((currentTime - rsISR.tLastCheck) >= 2)" by "if ((currentTime - rsISR.tLastCheck) >= 2000)". The reason for this change is that the millis() timer with slow (<16Mhz) Xtals is not increasing fast enough to obtain correct values.
On some boards this will make the program a bit slower, on other boards it may make the program a bit faster. If that becomes a problem, and Timer 1 is still available, the variant that uses Timer 1 should be selected.