For DxCore with < 48 pins, we now use RTC (requires PA0 is RS-RX)
For DxCore processors with less than 48 pins we now use RTC as default. Note that this requires the RS-Bus RX pin to be connected to PA0.
In addition, the library now uses TCB3 / RTC also for the DD, EA and EB series.
Probably this would also be possible with ATtiny 2/1/0-series.